Fix valocidad inicial spaceship

+++ tp4/tp4_java/src/main/java/ar/edu/itba/simulacion/tp4/marsMission/SimulationSettings.java
@@ -63,7 +63,7 @@ public final class SimulationSettings {
     @Jacksonized
     @Builder(setterPrefix = "with")
     public static class MarsMissionConfig {
-        public double               dt;
+        public int               dt;
         public double               gravitationalConstant;
         public MarsMissionSimulation.SpaceshipInitParams spaceship;
         public CelestialBodyData    sun;
